pyproject.toml
[project]
name = "vid2skill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "click>=8.1.8",
    "dill>=0.4.0",
    "drake>=1.40.0",
    "gymnasium>=1.1.1",
    "hydra-core>=1.3.2",
    "ipdb>=0.13.13",
    "scipy>=1.15.2",
]
